backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
alembic==1.5.8
    # via
    #   -r requirements.in
    #   ziggurat-foundations
amqp==5.0.6
    # via kombu
apispec[yaml]==4.4.0
    # via pyramid-apispec
authomatic==1.0.0
    # via -r requirements.in
backcall==0.2.0
    # via ipython
beaker==1.11.0
    # via pyramid-beaker
billiard==3.6.4.0
    # via celery
celery==5.0.5
    # via -r requirements.in
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
cryptography==3.4.7
    # via -r requirements.in
datadog==0.41.0
    # via -r requirements.in
decorator==5.0.7
    # via
    #   datadog
    #   dogpile.cache
    #   ipython
dnspython==2.1.0
    # via email-validator
dogpile.cache==1.1.2
    # via -r requirements.in
email-validator==1.1.2
    # via -r requirements.in
greenlet==1.0.0
    # via sqlalchemy
hupper==1.10.2
    # via pyramid
idna==2.10
    # via
    #   email-validator
    #   requests
ipython-genutils==0.2.0
    # via traitlets
ipython==7.22.0
    # via pyramid-ipython
jedi==0.18.0
    # via ipython
jinja2==2.11.3
    # via pyramid-jinja2
kombu==5.0.2
    # via celery
mako==1.1.4
    # via
    #   alembic
    #   pyramid-mako
markupsafe==1.1.1
    # via
    #   jinja2
    #   mako
    #   pyramid-jinja2
    #   webhelpers2
    #   wtforms
marshmallow==3.11.1
    # via -r requirements.in
paginate-sqlalchemy==0.3.1
    # via
    #   -r requirements.in
    #   ziggurat-foundations
paginate==0.5.6
    # via
    #   -r requirements.in
    #   paginate-sqlalchemy
    #   ziggurat-foundations
parso==0.8.2
    # via jedi
passlib==1.7.4
    # via ziggurat-foundations
pastedeploy==2.1.1
    # via plaster-pastedeploy
pbr==5.5.1
    # via stevedore
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
plaster-pastedeploy==0.7
    # via pyramid
plaster==1.0
    # via
    #   plaster-pastedeploy
    #   pyramid
prompt-toolkit==3.0.18
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.8.6
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pycparser==2.20
    # via cffi
pygments==2.8.1
    # via
    #   ipython
    #   pyramid-debugtoolbar
pyramid-apispec==0.4
    # via -r requirements.in
pyramid-beaker==0.8
    # via -r requirements.in
pyramid-debugtoolbar==4.9
    # via -r requirements.in
pyramid-ipython==0.2
    # via -r requirements.in
pyramid-jinja2==2.8
    # via -r requirements.in
pyramid-mailer==0.15.1
    # via -r requirements.in
pyramid-mako==1.1.0
    # via pyramid-debugtoolbar
pyramid-tm==2.4
    # via -r requirements.in
pyramid==2.0
    # via
    #   -r requirements.in
    #   pyramid-beaker
    #   pyramid-debugtoolbar
    #   pyramid-ipython
    #   pyramid-jinja2
    #   pyramid-mailer
    #   pyramid-mako
    #   pyramid-tm
python-dateutil==2.8.1
    # via alembic
python-editor==1.0.4
    # via alembic
pytz==2021.1
    # via celery
pyyaml==5.4.1
    # via apispec
redis==3.5.3
    # via -r requirements.in
repoze.lru==0.7
    # via pyramid-debugtoolbar
repoze.sendmail==4.4.1
    # via
    #   -r requirements.in
    #   pyramid-mailer
requests==2.25.1
    # via
    #   -r requirements.in
    #   datadog
sentry-sdk==1.0.0
    # via -r requirements.in
six==1.15.0
    # via
    #   click-repl
    #   python-dateutil
    #   webhelpers2
    #   ziggurat-foundations
sqlalchemy==1.4.9
    # via
    #   -r requirements.in
    #   alembic
    #   paginate-sqlalchemy
    #   ziggurat-foundations
    #   zope.sqlalchemy
stevedore==3.3.0
    # via dogpile.cache
traitlets==5.0.5
    # via ipython
transaction==3.0.1
    # via
    #   -r requirements.in
    #   pyramid-mailer
    #   pyramid-tm
    #   repoze.sendmail
    #   zope.sqlalchemy
translationstring==1.4
    # via pyramid
urllib3==1.26.4
    # via
    #   requests
    #   sentry-sdk
uwsgi==2.0.19.1
    # via -r requirements.in
venusian==3.0.0
    # via pyramid
vine==5.0.0
    # via
    #   amqp
    #   celery
waitress==2.0.0
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
webhelpers2-grid==0.9
    # via -r requirements.in
webhelpers2==2.0
    # via webhelpers2-grid
webob==1.8.7
    # via pyramid
wtforms==2.3.3
    # via -r requirements.in
ziggurat-foundations==0.8.4
    # via -r requirements.in
zope.deprecation==4.4.0
    # via
    #   pyramid
    #   pyramid-jinja2
    #   ziggurat-foundations
zope.interface==5.4.0
    # via
    #   pyramid
    #   repoze.sendmail
    #   transaction
    #   zope.sqlalchemy
zope.sqlalchemy==1.3
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
